Move Kafka message queue to Apache Kafka provider
