+++ providers/apache/kafka/provider.yaml
@@ -25,6 +25,7 @@ description: |
   `Apache Kafka  <https://kafka.apache.org/>`__
 # note that those versions are maintained by release manager - do not update them manually
 versions:
+  - 1.9.0
   - 1.8.1
   - 1.8.0
   - 1.7.0
@@ -74,3 +75,6 @@ triggers:
 connection-types:
   - hook-class-name: airflow.providers.apache.kafka.hooks.base.KafkaBaseHook
     connection-type: kafka
+
+queues:
+  - airflow.providers.apache.kafka.queues.kafka.KafkaMessageQueueProvider
