+++ providers/apache/kafka/tests/unit/apache/kafka/triggers/test_await_message.py
@@ -26,6 +26,8 @@ from airflow.providers.apache.kafka.hooks.consume import KafkaConsumerHook
 from airflow.providers.apache.kafka.triggers.await_message import AwaitMessageTrigger
 from airflow.utils import db
 
+from tests_common.test_utils.version_compat import AIRFLOW_V_3_0_PLUS, get_base_airflow_version_tuple
+
 pytestmark = pytest.mark.db_test
 
 
@@ -127,3 +129,16 @@ class TestTrigger:
         await asyncio.sleep(1.0)
         assert task.done() is False
         asyncio.get_event_loop().stop()
+
+
+@pytest.mark.skipif(not AIRFLOW_V_3_0_PLUS, reason="Requires Airflow 3.0.+")
+class TestMessageQueueTrigger:
+    def test_provider_integrations(self, cleanup_providers_manager):
+        if get_base_airflow_version_tuple() < (3, 0, 1):
+            pytest.skip("This test is only for Airflow 3.0.1+")
+
+    queue = "kafka://localhost:9092/topic1"
+    from airflow.providers.common.messaging.triggers.msg_queue import MessageQueueTrigger
+
+    trigger = MessageQueueTrigger(queue=queue)
+    assert isinstance(trigger.trigger, AwaitMessageTrigger)
